# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/tips:
    post:
      operationId: createTip
      tags: [admin]
      summary: Create a tip (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTipBody"
      responses:
        "201":
          description: Created
  /admin/tips/{id}:
    patch:
      operationId: updateTip
      tags: [admin]
      summary: Update a tip (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTipBody"
      responses:
        "200":
          description: Updated
    delete:
      operationId: deleteTip
      tags: [admin]
      summary: Delete a tip (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    CreateTipBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
      required:
        - title
        - description
        - category
    UpdateTipBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
...